Images are now just numbered.
Send Message nodes should send the number of the scene you want it to go to.

I don't know if you want the Wait to happen before or after a new scene image is loaded. At the moment it's before.

Dialoguer should have 4 scenes, Scene0, 1, 2 and 3. If it doesn't, import XML "18042014 (4)" in the Backup folder.